Split tuples, arrays and the C API off from the records file and refactored them
